api-key-service/requirement.txt
accelerate==1.5.2
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
bcrypt==5.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastapi==0.116.2
filelock==3.18.0
fonttools==4.61.1
fsspec==2025.3.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.3.2
parso==0.8.5
passlib==1.7.4
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.16.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tex==1.8
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.1
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
wcwidth==0.2.14
